Rename export_ to export- and remove unneeded .h files.
