+++ freud/box/export-Box.cc
@@ -1,7 +1,7 @@
 // Copyright (c) 2010-2024 The Regents of the University of Michigan
 // This file is from the freud project, released under the BSD 3-Clause License.
 
-#include "export_Box.h"
+#include "export-Box.h"
 
 namespace nb = nanobind;
