+++ freud/locality/export-NeighborQuery.cc
@@ -1,13 +1,18 @@
 // Copyright (c) 2010-2024 The Regents of the University of Michigan
 // This file is from the freud project, released under the BSD 3-Clause License.
 
-#include "export_NeighborQuery.h"
-
 #include <nanobind/nanobind.h>
 #include <nanobind/ndarray.h>
 #include <nanobind/stl/shared_ptr.h>
 
+#include "AABBQuery.h"
+#include "LinkCell.h"
+#include "RawPoints.h"
+
 namespace nb = nanobind;
+
+template<typename T, typename shape>
+using nb_array = nanobind::ndarray<T, shape, nanobind::device::cpu, nanobind::c_contig>;
 
 namespace freud { namespace locality {
