+++ freud/locality/export-PeriodicBuffer.cc
@@ -1,10 +1,10 @@
 // Copyright (c) 2010-2024 The Regents of the University of Michigan
 // This file is from the freud project, released under the BSD 3-Clause License.
 
-#include "export_PeriodicBuffer.h"
-
 #include <nanobind/stl/array.h>
 #include <nanobind/stl/shared_ptr.h>
+
+#include "PeriodicBuffer.h"
 
 namespace nb = nanobind;
