+++ freud/util/export-ManagedArray.h
